+++ packages/sui-dashboard/bin/sui-dashboard-components.js
@@ -53,8 +53,11 @@ const repositories = [
   'frontend-hab--web-app',
 
   'frontend-ij--uilib-widgets',
+  'frontend-ij--web-app',
 
-  'frontend-if--uilib-widgets'
+  'frontend-if--uilib-widgets',
+
+  'frontend-re--ut-web-app'
 ]
 
 const cloneSUIComponentsCommand = [