+++ packages/sui-js/src/string/to-query-string.js
@@ -6,12 +6,16 @@ import {stringify} from 'qs'
  * @param {object} queryParams
  * @param {object} [options={}]
  * @param {'indices'|'brackets'|'repeat'|'comma'} [options.arrayFormat] - specify the format of the output array
+ * @param {string} [options.delimiter] - delimiter
+ * @param {boolean} [options.encode=true] - encode
  */
 function toQueryString(queryParams, options = {}) {
-  const {arrayFormat} = options
+  const {arrayFormat, delimiter, encode = true} = options
 
   const mergedOptions = {
-    ...(typeof arrayFormat !== 'undefined' && {arrayFormat})
+    encode,
+    ...(typeof arrayFormat !== 'undefined' && {arrayFormat}),
+    ...(typeof delimiter !== 'undefined' && {delimiter})
   }
 
   return stringify(queryParams, mergedOptions)