+++ packages/sui-studio/src/components/test/index.js
@@ -2,15 +2,15 @@ import React, {useState, useEffect} from 'react'
 import PropTypes from 'prop-types'
 
 import cx from 'classnames'
+import hoistNonReactStatics from 'hoist-non-react-statics'
 
 import SUIContext from '@s-ui/react-context'
 import withContext from '../demo/HoC/withContext'
 import {cleanDisplayName} from '../demo/utilities'
 
 const BASE_CLASSNAME = 'sui-Test'
-const interOP = obj => (obj.default ? obj.default : obj)
 
-const Test = ({open, importTest, importComponent, context}) => {
+const Test = ({open, importTest, importComponent, contexts}) => {
   const [failures, setFailures] = useState(0)
   const [notFoundTest, setNotFoundTest] = useState(false)
 
@@ -23,21 +23,27 @@ const Test = ({open, importTest, importComponent, context}) => {
     importComponent().then(async module => {
       const Component = module.default || module
 
-      let nextContext = interOP(context)
-      nextContext =
-        typeof nextContext !== 'function' ? nextContext : await nextContext()
+      const nextContexts =
+        typeof contexts !== 'function' ? contexts : await contexts()
 
-      window.__STUDIO_CONTEXTS__ = nextContext
+      window.__STUDIO_CONTEXTS__ = nextContexts
       window.__STUDIO_COMPONENT__ = Component
 
-      const EnhanceComponent = withContext(nextContext, nextContext)(Component)
+      const EnhanceComponent = withContext(
+        nextContexts.default,
+        nextContexts
+      )(Component)
       !EnhanceComponent.displayName &&
         console.error('[sui-Test] Component without displayName') // eslint-disable-line
-      window[cleanDisplayName(EnhanceComponent.displayName)] = props => (
-        <SUIContext.Provider value={nextContext.default}>
+
+      const NextComponent = props => (
+        <SUIContext.Provider value={nextContexts.default}>
           <EnhanceComponent {...props} />
         </SUIContext.Provider>
       )
+      hoistNonReactStatics(NextComponent, Component)
+
+      window[cleanDisplayName(EnhanceComponent.displayName)] = NextComponent
 
       importTest()
         .then(() => {
@@ -49,7 +55,7 @@ const Test = ({open, importTest, importComponent, context}) => {
           setNotFoundTest(true)
         })
     })
-  }, [context, importComponent, importTest])
+  }, [contexts, importComponent, importTest])
 
   if (notFoundTest) {
     return <h1>Not found test</h1>
@@ -64,7 +70,7 @@ const Test = ({open, importTest, importComponent, context}) => {
 
 Test.displayName = 'Test'
 Test.propTypes = {
-  context: PropTypes.object,
+  contexts: PropTypes.object,
   importComponent: PropTypes.func,
   importTest: PropTypes.func,
   open: PropTypes.bool